// src/app/models/billing/cc-payment-model.ts
export class CCPaymentModel {
    
    token:string;
    charges: number;
    //cc_number: string;
    //cc_expiration_mm: string;
    //cc_expiration_yyyy: string;
    //cvc: string;
    card_holder_name: string;
    street_address1: string;
    street_address2: string;
    city: string;
    state: string;
    zip: string;
    service_name:string;
}